* Move accessory files to background directory
