+++ src/core/background/audio.js
