+++ src/core/background/get-named.js
