+++ src/core/background/pages.js
